Make library functions more flexible
Take Series (or maybe arrays...) rather then whole DataFrames where it
makes sense to do so, and make some of the interfaces more consistent.
Also make functions less dependent on the global configuration.

This should make the package easier to use as a library.

Does not cover 2D classification, but the next couple of commits
should...